requirements.txt
pillow
numpy>=1.13.3
aglfn>=0.0.4
